requirements.txt
cs50==9.3.0
Flask==3.0.0
Pillow==10.1.0
Pillow==10.1.0
Werkzeug==3.0.1
